app/api/grpc_server.py
"""gRPC server configuration and service implementation."""

import asyncio
from typing import Optional

import grpc
from opentelemetry.instrumentation.grpc import GrpcInstrumentorServer

from app.config import settings
from app.utils.logging import get_logger

logger = get_logger(__name__)


class DocumentServiceServicer:
    """gRPC service implementation for document operations."""
    
    def __init__(self):
        """Initialize the document service."""
        self.logger = get_logger(self.__class__.__name__)
    
    # TODO: Implement gRPC methods when protobuf definitions are ready


def create_grpc_server() -> grpc.aio.Server:
    """Create and configure gRPC server."""
    # Instrument gRPC with OpenTelemetry
    GrpcInstrumentorServer().instrument()
    
    server = grpc.aio.server()
    
    # TODO: Add service to server when protobuf definitions are ready
    # document_pb2_grpc.add_DocumentServiceServicer_to_server(
    #     DocumentServiceServicer(), server
    # )
    
    logger.info("gRPC server created")
    return server